+++ src/main/java/com/grinder/service/implement/SchedulerServiceImpl.java
@@ -13,6 +13,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.JobParametersBuilder;
 import org.springframework.batch.core.launch.JobLauncher;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
@@ -31,19 +32,17 @@ public class SchedulerServiceImpl implements SchedulerService {
     private final FeedRepository feedRepository;
     private final HeartRepository heartRepository;
     private final CommentRepository commentRepository;
-    private final JobLauncher jobLauncher;
-    private final Job logJob;
-
     private static final List<String> logList = new ArrayList<>();
 
-    //매일 0시에 전체 카페의 1/7만 평균 별점 업데이트 진행
-    @Override
+    // 매일 0시에 전체 카페의 1/7만 평균 별점 업데이트 진행
+
     @Scheduled(cron = "0 0 0 * * ?", zone = "Asia/Seoul")
     public void CalAverage() {
         executeWithRetry(this::performCalAverageTask, "CalAverage");
     }
 
-    private void performCalAverageTask() {
+    @Override
+    public void performCalAverageTask() {
         LocalDate today = LocalDate.now();
         int dayOfWeek = today.getDayOfWeek().getValue();
         updateAverageGradeForCafes(dayOfWeek);
@@ -55,21 +54,21 @@ public class SchedulerServiceImpl implements SchedulerService {
 
     @Override
     public void updateAverageGradeForCafes(int dayOfWeek) {
-        // 전체 카페 수 조회
         long totalCafes = cafeRepository.count();
-        long batchSize = totalCafes / 7;  // 한 번에 처리할 카페 수
-        long offset = (dayOfWeek - 1) * batchSize;  // 주중의 날에 따라 오프셋 조정
+        long batchSize = totalCafes / 7;
+        long offset = (dayOfWeek - 1) * batchSize;
 
-        // 해당 요일에 처리할 카페 리스트 가져오기
         List<Cafe> cafes = cafeRepository.findCafesForAverageCalculation(batchSize, offset);
         cafes.forEach(this::calculateAndSetAverageGrade);
     }
+
     @Override
     public void calculateAndSetAverageGrade(Cafe cafe) {
         Double average = cafeRepository.findAverageGradeByCafeId(cafe.getCafeId());
         cafe.setAverageGrade(average.intValue());
         cafeRepository.save(cafe);
     }
+
     @Override
     public void updateTagListForMembers(int dayOfWeek) {
         long totalMembers = memberRepository.count();
@@ -86,7 +85,8 @@ public class SchedulerServiceImpl implements SchedulerService {
         executeWithRetry(this::performRecommendAlanTask, "recommendAlan");
     }
 
-    private void performRecommendAlanTask() {
+    @Override
+    public void performRecommendAlanTask() {
         LocalDate today = LocalDate.now();
         int dayOfMonth = today.getDayOfMonth();
         recommendCafeForMembers(dayOfMonth);
@@ -96,8 +96,8 @@ public class SchedulerServiceImpl implements SchedulerService {
     @Override
     public void recommendCafeForMembers(int dayOfMonth) {
         long totalMembers = memberRepository.count();
-        long batchSize = totalMembers / 30;  // 전체 회원을 30으로 나누어 하루 처리할 회원 수 계산
-        long offset = (dayOfMonth - 1) * batchSize;  // 일자에 따른 오프셋 계산
+        long batchSize = totalMembers / 30;
+        long offset = (dayOfMonth - 1) * batchSize;
 
         List<Member> members = memberRepository.findMembersForTagUpdate(batchSize, offset);
         for (Member member : members) {
@@ -108,8 +108,8 @@ public class SchedulerServiceImpl implements SchedulerService {
     @Override
     public void updateRanks(int dayOfMonth) {
         long totalMembers = feedRepository.count();
-        long batchSize = totalMembers / 30;  // 전체 회원을 30으로 나누어 하루 처리할 회원 수 계산
-        long offset = (dayOfMonth - 1) * batchSize;  // 일자에 따른 오프셋 계산
+        long batchSize = totalMembers / 30;
+        long offset = (dayOfMonth - 1) * batchSize;
 
         List<Feed> feeds = feedRepository.findFeedsForRankUpdate(batchSize, offset);
         for (Feed feed : feeds) {
@@ -118,13 +118,15 @@ public class SchedulerServiceImpl implements SchedulerService {
             feedRepository.updateFeedRank(feed.getFeedId(), rank.intValue());
         }
     }
+
     @Override
     @Scheduled(cron = "0 0 2 * * ?", zone = "Asia/Seoul")
     public void updateRank() {
         executeWithRetry(this::performUpdateRankTask, "updateRank");
     }
 
-    private void performUpdateRankTask() {
+    @Override
+    public void performUpdateRankTask() {
         LocalDate today = LocalDate.now();
         int dayOfMonth = today.getDayOfMonth();
         updateRanks(dayOfMonth);
@@ -153,20 +155,9 @@ public class SchedulerServiceImpl implements SchedulerService {
     public List<String> getLogList() {
         return new ArrayList<>(logList);
     }
+
     @Override
     public void clearLogList() {
         logList.clear();
-    }
-
-    @Override
-    @Scheduled(cron = "0 0 3 * * ?", zone = "Asia/Seoul")
-    public void runLogJob() {
-        try {
-            jobLauncher.run(logJob, new JobParametersBuilder()
-                    .addLong("time", System.currentTimeMillis())
-                    .toJobParameters());
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
     }
 }
