+++ src/test/java/com/grinder/GrinderApplicationTests.java
@@ -1,10 +1,12 @@
 package com.grinder;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
+@EnableBatchProcessing
 @ActiveProfiles("test")
 class GrinderApplicationTests {
